restructure tests folder
