+++ tests/domain/test_aminoacidlist.py
