+++ tests/utils/test_graph.py
@@ -6,8 +6,8 @@ from pdb2sql import pdb2sql
 import numpy as np
 from deeprankcore.utils.grid import GridSettings, MapMethod
 from deeprankcore.utils.graph import Graph, Edge, Node
-from deeprankcore.molstruct.pair import ResidueContact
 from deeprankcore.utils.buildgraph import get_structure
+from deeprankcore.molstruct.pair import ResidueContact
 from deeprankcore.domain import (nodefeatures as Nfeat,
                                 edgefeatures as Efeat)
